Phase 3 foundation: Memory64 type widening and dual ABI dispatch
Enable WASM_ENABLE_MEMORY64=1 alongside WASM_ENABLE_FAST_INTERP=1 (cmake
rejects this combo but the C code compiles fine — Memory64 is AOT-only).

Type widening:
- HlWasmCallOpts: max_input/max_output uint32_t → uint64_t
- HlWasmInstance: default_max_input/output uint32_t → uint64_t
- HlRuntime wasm_config: max_input/output uint32_t → uint64_t
- HlWasmDataSegment: wasm_addr uint32_t → uint64_t
- HL_WASM64_MAX_IO_SIZE: 16 GB (vs 256 MB for WASM32)

Dual ABI dispatch in hl_cap_wasm_call_buf and instance_call_buf:
- WASM32: hull_process(i32,i32,i32,i32) → 4 argv cells
- Memory64: hull_process(i64,i64,i64,i64) → 8 argv cells (2 per i64)

Memory64 detection via WAMR internal WASMMemoryInstance.is_memory64,
stored as HlWasmModule.is_memory64. Non-AOT Memory64 modules rejected
with "memory64_requires_aot" error.

Source-compatible: existing uint32 callers assign to uint64 fields without
casts. All WASM32 tests pass unchanged.

+++ include/hull/runtime.h
@@ -59,8 +59,8 @@ struct HlRuntime {
         uint32_t heap_size;   /* ceiling: 0 = use compile-time default */
         uint32_t stack_size;
         int64_t  gas;
-        uint32_t max_input;
-        uint32_t max_output;
+        uint64_t max_input;
+        uint64_t max_output;
     } wasm_config;                     /* three-tier resolved limits (CLI > manifest > defaults) */
 #endif
 };
